refactor(Piece): 공통 부분 메서드 Piece로 공통화
