+++ src/main/java/janggi/domain/piece/Cannon.java
@@ -5,7 +5,6 @@ import janggi.domain.Team;
 import janggi.domain.piece.direction.Direction;
 import janggi.domain.piece.direction.Position;
 import janggi.domain.piece.direction.Route;
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -26,20 +25,6 @@ public class Cannon extends Piece {
             validRoutes.addAll(generateRoutesInDirection(direction));
         }
         return validRoutes;
-    }
-
-    private Set<Route> generateRoutesInDirection(final Direction direction) {
-        final Set<Route> directionalRoutes = new HashSet<>();
-
-        Position currentPosition = position;
-        final List<Position> positions = new ArrayList<>();
-        while (currentPosition.canMove(direction)) {
-            final Position nextPosition = currentPosition.move(direction);
-            positions.add(nextPosition);
-            directionalRoutes.add(new Route(new ArrayList<>(positions)));
-            currentPosition = nextPosition;
-        }
-        return directionalRoutes;
     }
 
     @Override
