+++ src/main/java/janggi/domain/piece/Elephant.java
@@ -12,12 +12,7 @@ import static janggi.domain.piece.direction.Direction.UP;
 import janggi.domain.Team;
 import janggi.domain.piece.direction.Direction;
 import janggi.domain.piece.direction.Position;
-import janggi.domain.piece.direction.Route;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class Elephant extends Piece {
 
@@ -40,27 +35,7 @@ public class Elephant extends Piece {
     }
 
     @Override
-    public Set<Route> calculateIndependentRoutes() {
-        return ELEPHANT_MOVES.stream()
-                .map(this::calculateRoute)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .collect(Collectors.toSet());
-    }
-
-    private Optional<Route> calculateRoute(final List<Direction> move) {
-        final List<Position> positions = new ArrayList<>();
-
-        Position currentPosition = position;
-        for (final Direction direction : move) {
-            if (currentPosition.canMove(direction)) {
-                final Position nextPosition = currentPosition.move(direction);
-                positions.add(nextPosition);
-                currentPosition = nextPosition;
-                continue;
-            }
-            return Optional.empty();
-        }
-        return Optional.of(new Route(positions));
+    protected List<List<Direction>> getMoveStrategy() {
+        return ELEPHANT_MOVES;
     }
 }
